oddgnome
This one was difficult for me. The biggest hangup was
trying to break when the out-of-order number was found
without advancing the scanf; I used a found flag in the
end. This may be a use case for fgets, where you read
the line all at once and sscanf each number out.